# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Bracket Boss API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type TournamentStatus = typeof TournamentStatus[keyof typeof TournamentStatus];


export const TournamentStatus = {
  lobby: 'lobby',
  active: 'active',
  completed: 'completed',
} as const;

export interface Tournament {
  id: string;
  name: string;
  status: TournamentStatus;
  registrationLocked?: boolean;
  createdAt: string;
}

export type TournamentWithTokenStatus = typeof TournamentWithTokenStatus[keyof typeof TournamentWithTokenStatus];


export const TournamentWithTokenStatus = {
  lobby: 'lobby',
  active: 'active',
  completed: 'completed',
} as const;

export interface TournamentWithToken {
  id: string;
  name: string;
  status: TournamentWithTokenStatus;
  registrationLocked?: boolean;
  createdAt: string;
  hostToken: string;
}

export type TournamentFullStatus = typeof TournamentFullStatus[keyof typeof TournamentFullStatus];


export const TournamentFullStatus = {
  lobby: 'lobby',
  active: 'active',
  completed: 'completed',
} as const;

export interface Player {
  id: string;
  tournamentId: string;
  firstName: string;
  lastName: string;
  /** @nullable */
  partnerName?: string | null;
  /** @nullable */
  teamName?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  eloRating: number;
  /** @nullable */
  rankTitle?: string | null;
  /** @nullable */
  rankEmoji?: string | null;
  seed: number;
  joinedAt: string;
}

export type MatchBracket = typeof MatchBracket[keyof typeof MatchBracket];


export const MatchBracket = {
  winner: 'winner',
  loser: 'loser',
  grand_finals: 'grand_finals',
  grand_finals_reset: 'grand_finals_reset',
} as const;

export type MatchStatus = typeof MatchStatus[keyof typeof MatchStatus];


export const MatchStatus = {
  pending: 'pending',
  active: 'active',
  completed: 'completed',
  bye: 'bye',
} as const;

export interface Match {
  id: string;
  tournamentId: string;
  round: number;
  matchNumber: number;
  bracket: MatchBracket;
  /** @nullable */
  playerOneId?: string | null;
  /** @nullable */
  playerTwoId?: string | null;
  /** @nullable */
  winnerId?: string | null;
  /** @nullable */
  scoreOne?: number | null;
  /** @nullable */
  scoreTwo?: number | null;
  status: MatchStatus;
  isBye: boolean;
  /** @nullable */
  nextWinnerMatchId?: string | null;
  /** @nullable */
  nextWinnerSlot?: string | null;
  /** @nullable */
  nextLoserMatchId?: string | null;
  /** @nullable */
  nextLoserSlot?: string | null;
  /** @nullable */
  completedAt?: string | null;
}

export interface TournamentFull {
  id: string;
  name: string;
  status: TournamentFullStatus;
  registrationLocked?: boolean;
  createdAt: string;
  /** @nullable */
  startedAt?: string | null;
  /** @nullable */
  completedAt?: string | null;
  players: Player[];
  matches: Match[];
}

export interface TournamentSummary {
  tournamentId: string;
  champion: Player;
  runnerUp: Player;
  thirdPlace?: Player | null;
  totalMatches: number;
  playerCount: number;
  /** @nullable */
  durationMinutes?: number | null;
}

export interface PlayerWithToken {
  id: string;
  tournamentId: string;
  firstName: string;
  lastName: string;
  /** @nullable */
  partnerName?: string | null;
  /** @nullable */
  teamName?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  eloRating: number;
  /** @nullable */
  rankTitle?: string | null;
  /** @nullable */
  rankEmoji?: string | null;
  seed: number;
  joinedAt: string;
  playerToken: string;
}

export interface MatchHistoryEntry {
  matchId: string;
  tournamentId: string;
  tournamentName: string;
  bracket: string;
  round: number;
  opponentName: string;
  won: boolean;
  /** @nullable */
  scoreOne?: number | null;
  /** @nullable */
  scoreTwo?: number | null;
  completedAt: string;
}

export interface PlayerStats {
  player: Player;
  wins: number;
  losses: number;
  matchesPlayed: number;
  winPct: number;
  tournamentWins: number;
  recentMatches: MatchHistoryEntry[];
}

export interface OpenPlayMatchRecord {
  id: string;
  winnerTeam: number;
  /** @nullable */
  scoreOne?: number | null;
  /** @nullable */
  scoreTwo?: number | null;
  teamOnePlayers?: Player[];
  teamTwoPlayers?: Player[];
  playedAt: string;
}

export interface OpenPlayState {
  pool: Player[];
  recentMatches: OpenPlayMatchRecord[];
}

export interface OpenPlayMatchInput {
  hostToken: string;
  teamOnePOneId: string;
  teamOnePTwoId?: string;
  teamTwoPOneId: string;
  teamTwoPTwoId?: string;
  winnerTeam: number;
  scoreOne?: number;
  scoreTwo?: number;
}

export interface TournamentInput {
  name?: string;
}

export interface TournamentUpdate {
  name?: string;
  registrationLocked?: boolean;
  hostToken?: string;
}

export interface PlayerInput {
  /** @minLength 1 */
  firstName: string;
  /** @minLength 1 */
  lastName: string;
  partnerName?: string;
  teamName?: string;
}

export interface PlayerUpdate {
  playerToken: string;
  teamName?: string;
  avatarUrl?: string;
}

export interface MatchUpdate {
  hostToken: string;
  winnerId?: string;
  scoreOne?: number;
  scoreTwo?: number;
}

export interface HostTokenInput {
  hostToken: string;
}

export interface RequestUploadUrlBody {
  name: string;
  size: number;
  contentType: string;
}

export interface RequestUploadUrlResponse {
  uploadURL: string;
  objectPath: string;
}