# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Bracket Boss API
servers:
  - url: /api
...
paths:
...
  /tournaments/{tournamentId}/matches/{matchId}:
    patch:
      operationId: updateMatch
      tags: [matches]
      parameters:
        - name: tournamentId
          in: path
          required: true
          schema:
            type: string
        - name: matchId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MatchUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TournamentFull"
  /tournaments/{tournamentId}/matches/undo:
    post:
      operationId: undoLastMatch
      tags: [matches]
      parameters:
        - name: tournamentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HostTokenInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TournamentFull"
...
  /tournaments/{tournamentId}/open-play:
    get:
      operationId: getOpenPlayPool
      tags: [openplay]
      parameters:
        - name: tournamentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenPlayState"
  /tournaments/{tournamentId}/open-play/matches:
    post:
      operationId: logOpenPlayMatch
      tags: [openplay]
      parameters:
        - name: tournamentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenPlayMatchInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenPlayState"
...
components:
  schemas:
...
    TournamentFull:
      type: object
      required: [id, name, status, createdAt, players, matches]
      properties:
        id:
          type: string
        name:
          type: string
        status:
          type: string
          enum: [lobby, active, completed]
        registrationLocked:
          type: boolean
        createdAt:
          type: string
          format: date-time
        startedAt:
          type: ["string", "null"]
          format: date-time
        completedAt:
          type: ["string", "null"]
          format: date-time
        players:
          type: array
          items:
            $ref: "#/components/schemas/Player"
        matches:
          type: array
          items:
            $ref: "#/components/schemas/Match"
...
    Player:
      type: object
      required: [id, tournamentId, firstName, lastName, seed, joinedAt, eloRating]
      properties:
        id:
          type: string
        tournamentId:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        partnerName:
          type: ["string", "null"]
        teamName:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        eloRating:
          type: number
        rankTitle:
          type: ["string", "null"]
        rankEmoji:
          type: ["string", "null"]
        seed:
          type: integer
        joinedAt:
          type: string
          format: date-time
...
    Match:
      type: object
      required: [id, tournamentId, round, matchNumber, bracket, status, isBye]
      properties:
        id:
          type: string
        tournamentId:
          type: string
        round:
          type: integer
        matchNumber:
          type: integer
        bracket:
          type: string
          enum: [winner, loser, grand_finals, grand_finals_reset]
        playerOneId:
          type: ["string", "null"]
        playerTwoId:
          type: ["string", "null"]
        winnerId:
          type: ["string", "null"]
        scoreOne:
          type: ["integer", "null"]
        scoreTwo:
          type: ["integer", "null"]
        status:
          type: string
          enum: [pending, active, completed, bye]
        isBye:
          type: boolean
        nextWinnerMatchId:
          type: ["string", "null"]
        nextWinnerSlot:
          type: ["string", "null"]
        nextLoserMatchId:
          type: ["string", "null"]
        nextLoserSlot:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
          format: date-time
    OpenPlayState:
      type: object
      required: [pool, recentMatches]
      properties:
        pool:
          type: array
          items:
            $ref: "#/components/schemas/Player"
        recentMatches:
          type: array
          items:
            $ref: "#/components/schemas/OpenPlayMatchRecord"
    OpenPlayMatchRecord:
      type: object
      required: [id, winnerTeam, playedAt]
      properties:
        id:
          type: string
        winnerTeam:
          type: integer
        scoreOne:
          type: ["integer", "null"]
        scoreTwo:
          type: ["integer", "null"]
        teamOnePlayers:
          type: array
          items:
            $ref: "#/components/schemas/Player"
        teamTwoPlayers:
          type: array
          items:
            $ref: "#/components/schemas/Player"
        playedAt:
          type: string
          format: date-time
    OpenPlayMatchInput:
      type: object
      required: [hostToken, teamOnePOneId, teamTwoPOneId, winnerTeam]
      properties:
        hostToken:
          type: string
        teamOnePOneId:
          type: string
        teamOnePTwoId:
          type: string
        teamTwoPOneId:
          type: string
        teamTwoPTwoId:
          type: string
        winnerTeam:
          type: integer
        scoreOne:
          type: integer
        scoreTwo:
          type: integer
...
    MatchUpdate:
      type: object
      required: [hostToken]
      properties:
        hostToken:
          type: string
        winnerId:
          type: string
        scoreOne:
          type: integer
        scoreTwo:
          type: integer
    HostTokenInput:
      type: object
      required: [hostToken]
      properties:
        hostToken:
          type: string
...